requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Login
Flask-Bcrypt
gunicorn
requests
email-validator
Flask-WTF
Flask-Limiter
Flask-Talisman
user-agents
